refactor(plugins): modularize plugin system into separate crates
