+++ crates/auroraview-plugins/core/src/error.rs
@@ -1,47 +1,10 @@
-//! Plugin type definitions
+//! Plugin error types
 //!
-//! Common types and error handling for the plugin system.
+//! Common error handling for the plugin system.
 
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use thiserror::Error;
-
-/// Plugin command specification
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PluginCommand {
-    /// Command name
-    pub name: String,
-    /// Command description
-    pub description: String,
-    /// Required arguments
-    pub required_args: Vec<String>,
-    /// Optional arguments
-    pub optional_args: Vec<String>,
-}
-
-impl PluginCommand {
-    /// Create a new command specification
-    pub fn new(name: impl Into<String>, description: impl Into<String>) -> Self {
-        Self {
-            name: name.into(),
-            description: description.into(),
-            required_args: Vec::new(),
-            optional_args: Vec::new(),
-        }
-    }
-
-    /// Add required arguments
-    pub fn with_required(mut self, args: &[&str]) -> Self {
-        self.required_args = args.iter().map(|s| s.to_string()).collect();
-        self
-    }
-
-    /// Add optional arguments
-    pub fn with_optional(mut self, args: &[&str]) -> Self {
-        self.optional_args = args.iter().map(|s| s.to_string()).collect();
-        self
-    }
-}
 
 /// Plugin error codes
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
