+++ crates/auroraview-plugins/core/src/scope.rs
