+++ crates/auroraview-plugins/fs/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "auroraview-plugin-fs"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "AuroraView File System Plugin: Native file system operations for WebView applications"
+keywords = ["webview", "plugins", "filesystem", "io"]
+categories = ["gui", "filesystem"]
+
+[lib]
+name = "auroraview_plugin_fs"
+path = "src/lib.rs"
+
+[dependencies]
+# Core plugin framework
+auroraview-plugin-core = { path = "../core" }
+
+# Logging
+tracing = "0.1"
+
+# Serialization
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+# Binary encoding
+base64 = "0.22"
