+++ crates/auroraview-plugins/fs/src/types.rs
