+++ crates/auroraview-plugins/src/lib.rs
@@ -21,12 +21,19 @@
 //! └─────────────────────────────────────────────────────────────┘
 //! ```
 //!
+//! ## Crate Structure
+//!
+//! - `auroraview-plugin-core` - Core plugin framework (traits, router, scope)
+//! - `auroraview-plugin-fs` - File system plugin
+//! - `auroraview-plugins` - This crate, aggregates all plugins
+//!
 //! ## Available Plugins
 //!
 //! - **fs**: File system operations (read, write, list, etc.)
 //! - **clipboard**: System clipboard access (read/write text, images)
 //! - **shell**: Execute commands, open URLs/files
 //! - **dialog**: Native file/folder dialogs
+//! - **process**: Process spawning with IPC support
 //!
 //! ## Command Format
 //!
@@ -34,266 +41,44 @@
 //!
 //! Example: `plugin:fs|read_file`
 
+// Re-export core types
+pub use auroraview_plugin_core::{
+    PathScope, PluginCommand, PluginError, PluginErrorCode, PluginEventCallback, PluginHandler,
+    PluginRequest, PluginResponse, PluginResult, PluginRouter, ScopeConfig, ScopeError, ShellScope,
+};
+
+// Re-export fs plugin
+pub use auroraview_plugin_fs as fs;
+
+// Built-in plugins (still in this crate for now)
 pub mod clipboard;
 pub mod dialog;
-pub mod fs;
 pub mod process;
-pub mod scope;
 pub mod shell;
-mod types;
-// Note: assets module is disabled until JS files are created
-// mod assets;
 
-pub use scope::{PathScope, ScopeConfig, ScopeError, ShellScope};
-pub use types::{PluginCommand, PluginError, PluginErrorCode, PluginResult};
+use std::sync::Arc;
 
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use std::collections::HashMap;
-use std::sync::{Arc, RwLock};
+/// Create a plugin router with all built-in plugins registered
+pub fn create_router() -> PluginRouter {
+    let mut router = PluginRouter::new();
+    let event_callback = router.event_callback_ref();
 
-/// Event callback type for plugins to emit events
-pub type PluginEventCallback = Arc<dyn Fn(&str, Value) + Send + Sync>;
+    // Register built-in plugins
+    router.register("fs", Arc::new(fs::FsPlugin::new()));
+    router.register("clipboard", Arc::new(clipboard::ClipboardPlugin::new()));
+    router.register("shell", Arc::new(shell::ShellPlugin::new()));
+    router.register("dialog", Arc::new(dialog::DialogPlugin::new()));
 
-/// Plugin command request
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PluginRequest {
-    /// Plugin name (e.g., "fs", "clipboard")
-    pub plugin: String,
-    /// Command name (e.g., "read_file", "write")
-    pub command: String,
-    /// Command arguments as JSON
-    pub args: Value,
-    /// Optional request ID for async response
-    pub id: Option<String>,
+    // Create process plugin with shared event callback
+    let process_plugin = process::ProcessPlugin::with_event_callback(event_callback);
+    router.register("process", Arc::new(process_plugin));
+
+    router
 }
 
-impl PluginRequest {
-    /// Parse a command string in format "plugin:<plugin>|<command>"
-    pub fn from_invoke(invoke_cmd: &str, args: Value) -> Option<Self> {
-        if !invoke_cmd.starts_with("plugin:") {
-            return None;
-        }
-
-        let rest = &invoke_cmd[7..]; // Skip "plugin:"
-        let parts: Vec<&str> = rest.splitn(2, '|').collect();
-        if parts.len() != 2 {
-            return None;
-        }
-
-        Some(Self {
-            plugin: parts[0].to_string(),
-            command: parts[1].to_string(),
-            args,
-            id: None,
-        })
-    }
-
-    /// Create a new plugin request
-    pub fn new(plugin: impl Into<String>, command: impl Into<String>, args: Value) -> Self {
-        Self {
-            plugin: plugin.into(),
-            command: command.into(),
-            args,
-            id: None,
-        }
-    }
-
-    /// Set the request ID
-    pub fn with_id(mut self, id: impl Into<String>) -> Self {
-        self.id = Some(id.into());
-        self
-    }
-}
-
-/// Plugin command response
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PluginResponse {
-    /// Success flag
-    pub success: bool,
-    /// Response data (if success)
-    pub data: Option<Value>,
-    /// Error message (if failure)
-    pub error: Option<String>,
-    /// Error code (if failure)
-    pub code: Option<String>,
-    /// Request ID (echoed from request)
-    pub id: Option<String>,
-}
-
-impl PluginResponse {
-    /// Create a success response
-    pub fn ok(data: Value) -> Self {
-        Self {
-            success: true,
-            data: Some(data),
-            error: None,
-            code: None,
-            id: None,
-        }
-    }
-
-    /// Create an error response
-    pub fn err(error: impl Into<String>, code: impl Into<String>) -> Self {
-        Self {
-            success: false,
-            data: None,
-            error: Some(error.into()),
-            code: Some(code.into()),
-            id: None,
-        }
-    }
-
-    /// Set the request ID
-    pub fn with_id(mut self, id: Option<String>) -> Self {
-        self.id = id;
-        self
-    }
-}
-
-/// Trait for plugin implementations
-pub trait PluginHandler: Send + Sync {
-    /// Plugin name
-    fn name(&self) -> &str;
-
-    /// Handle a command
-    fn handle(&self, command: &str, args: Value, scope: &ScopeConfig) -> PluginResult<Value>;
-
-    /// Get supported commands
-    fn commands(&self) -> Vec<&str>;
-}
-
-/// Plugin router for dispatching commands to plugins
-pub struct PluginRouter {
-    /// Registered plugins
-    plugins: HashMap<String, Arc<dyn PluginHandler>>,
-    /// Global scope configuration
-    scope: ScopeConfig,
-    /// Event callback for plugins to emit events to frontend
-    event_callback: Arc<RwLock<Option<PluginEventCallback>>>,
-}
-
-impl Default for PluginRouter {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl PluginRouter {
-    /// Create a new plugin router with default plugins
-    pub fn new() -> Self {
-        let event_callback: Arc<RwLock<Option<PluginEventCallback>>> = Arc::new(RwLock::new(None));
-
-        let mut router = Self {
-            plugins: HashMap::new(),
-            scope: ScopeConfig::new(),
-            event_callback: event_callback.clone(),
-        };
-
-        // Register built-in plugins
-        router.register("fs", Arc::new(fs::FsPlugin::new()));
-        router.register("clipboard", Arc::new(clipboard::ClipboardPlugin::new()));
-        router.register("shell", Arc::new(shell::ShellPlugin::new()));
-        router.register("dialog", Arc::new(dialog::DialogPlugin::new()));
-
-        // Create process plugin with shared event callback
-        let process_plugin = process::ProcessPlugin::with_event_callback(event_callback);
-        router.register("process", Arc::new(process_plugin));
-
-        router
-    }
-
-    /// Create with custom scope configuration
-    pub fn with_scope(scope: ScopeConfig) -> Self {
-        let mut router = Self::new();
-        router.scope = scope;
-        router
-    }
-
-    /// Set the event callback for plugins to emit events
-    ///
-    /// This callback will be invoked when plugins (like ProcessPlugin) need
-    /// to send events to the frontend (e.g., process stdout/stderr output).
-    pub fn set_event_callback(&self, callback: PluginEventCallback) {
-        tracing::info!("[PluginRouter] Setting event callback");
-        let mut cb = self.event_callback.write().unwrap();
-        *cb = Some(callback);
-        tracing::info!("[PluginRouter] Event callback set successfully");
-    }
-
-    /// Clear the event callback
-    pub fn clear_event_callback(&self) {
-        let mut cb = self.event_callback.write().unwrap();
-        *cb = None;
-    }
-
-    /// Emit an event through the callback (if set)
-    pub fn emit_event(&self, event_name: &str, data: Value) {
-        if let Some(callback) = self.event_callback.read().unwrap().as_ref() {
-            callback(event_name, data);
-        }
-    }
-
-    /// Register a plugin
-    pub fn register(&mut self, name: impl Into<String>, plugin: Arc<dyn PluginHandler>) {
-        self.plugins.insert(name.into(), plugin);
-    }
-
-    /// Unregister a plugin
-    pub fn unregister(&mut self, name: &str) -> Option<Arc<dyn PluginHandler>> {
-        self.plugins.remove(name)
-    }
-
-    /// Handle a plugin command
-    pub fn handle(&self, request: PluginRequest) -> PluginResponse {
-        // Check if plugin is enabled
-        if !self.scope.is_plugin_enabled(&request.plugin) {
-            return PluginResponse::err(
-                format!("Plugin '{}' is disabled", request.plugin),
-                "PLUGIN_DISABLED",
-            )
-            .with_id(request.id);
-        }
-
-        let plugin = match self.plugins.get(&request.plugin) {
-            Some(p) => p,
-            None => {
-                return PluginResponse::err(
-                    format!("Plugin '{}' not found", request.plugin),
-                    "PLUGIN_NOT_FOUND",
-                )
-                .with_id(request.id);
-            }
-        };
-
-        match plugin.handle(&request.command, request.args.clone(), &self.scope) {
-            Ok(data) => PluginResponse::ok(data).with_id(request.id),
-            Err(e) => PluginResponse::err(e.message(), e.code()).with_id(request.id),
-        }
-    }
-
-    /// Check if a plugin is registered
-    pub fn has_plugin(&self, name: &str) -> bool {
-        self.plugins.contains_key(name)
-    }
-
-    /// Get list of registered plugin names
-    pub fn plugin_names(&self) -> Vec<&str> {
-        self.plugins.keys().map(|s| s.as_str()).collect()
-    }
-
-    /// Get the scope configuration
-    pub fn scope(&self) -> &ScopeConfig {
-        &self.scope
-    }
-
-    /// Get mutable scope configuration
-    pub fn scope_mut(&mut self) -> &mut ScopeConfig {
-        &mut self.scope
-    }
-
-    /// Update scope configuration
-    pub fn set_scope(&mut self, scope: ScopeConfig) {
-        self.scope = scope;
-    }
+/// Create a plugin router with custom scope configuration
+pub fn create_router_with_scope(scope: ScopeConfig) -> PluginRouter {
+    let mut router = create_router();
+    router.set_scope(scope);
+    router
 }
